feat: control IDCTA fetching with toggles
