+++ src/app/lib/utilities/getToggleDefinition/index.tsx
@@ -1,0 +1,10 @@
+import { ToggleDefinition, Toggles } from '#app/models/types/global';
+
+const getToggleDefinitions = (
+  toggles: Toggles = {},
+): Record<string, ToggleDefinition> => {
+  const { _environment, ...toggleDefinitions } = toggles;
+  return toggleDefinitions;
+};
+
+export default getToggleDefinitions;
