+++ src/app/utilities/createAdNonce/index.ts
@@ -1,5 +1,6 @@
 import getUUID from '#app/lib/utilities/getUUID';
-import { Toggles, ToggleDefinition } from '#app/models/types/global';
+import { Toggles } from '#app/models/types/global';
+import getToggleDefinitions from '#app/lib/utilities/getToggleDefinition';
 
 interface InjectNonceHeader {
   toggles: Toggles;
@@ -8,13 +9,6 @@ interface InjectNonceHeader {
   isLite: boolean;
   isAmp: boolean;
 }
-
-const getToggleDefinitions = (
-  toggles: Toggles = {},
-): Record<string, ToggleDefinition> => {
-  const { _environment, ...toggleDefinitions } = toggles;
-  return toggleDefinitions;
-};
 
 const isNonceEnabledForCountry = (
   nonceCountries: string | number | undefined,
